+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   instead of being interpreted as `stdin` (#351, @NathanReb)
 + Update to cmdliner 2 and install completion scripts for common shells
   (#351, @NathanReb)
++ Add explicit support for external types (#353, @NathanReb)
 
 ## 1.9.0